# MAINT: Speed-up random_polyagamma_fill function