--- a/src/pgm_random.c
+++ b/src/pgm_random.c
@@ -2,6 +2,7 @@
  *
  * SPDX-License-Identifier: BSD-3-Clause */
 #include <math.h>
+#include <string.h>
 #include "pgm_devroye.h"
 #include "pgm_alternate.h"
 #include "pgm_saddle.h"
@@ -13,20 +14,11 @@
     #define PGM_EXTERN
 #endif
 
-PGM_EXTERN PGM_INLINE void
-pgm_random_polyagamma_fill(bitgen_t* bitgen_state, double h, double z,
-                           sampler_t method, size_t n, double* out);
-
-PGM_EXTERN PGM_INLINE void
-pgm_random_polyagamma_fill2(bitgen_t* bitgen_state, const double* h,
-                            const double* z, sampler_t method, size_t n,
-                            double* restrict out);
-
 /* numpy c-api declarations */
 PGM_EXTERN double
-random_standard_normal(bitgen_t *bitgen_state);
+random_standard_normal(bitgen_t* bitgen_state);
 PGM_EXTERN double
-random_standard_gamma(bitgen_t *bitgen_state, double shape);
+random_standard_gamma(bitgen_t* bitgen_state, double shape);
 
 /*
  * Sample from a PG(h. z) using a Normal Approximation. For sufficiently large
@@ -38,21 +30,25 @@ random_standard_gamma(bitgen_t *bitgen_state, double shape);
  *   approaches 0. The formula can be easily generated using any online math
  *   equation calculator.
  */
-static PGM_INLINE double
-random_polyagamma_normal_approx(bitgen_t* bitgen_state, double h, double z)
+static PGM_INLINE void
+random_polyagamma_normal_approx(bitgen_t* bitgen_state, double h, double z,
+                                size_t n, double* out)
 {
-    double x, mean, variance;
+    double x, mean, stdev;
 
     if (z == 0.) {
         mean = 0.25 * h;
-        variance = 0.041666688 * h;
+        stdev = sqrt(0.041666688 * h);
     }
     else {
         x = tanh(0.5 * z);
         mean = 0.5 * h * x / z;
-        variance = 0.25 * h * (sinh(z) - z) * (1. - x * x) / (z * z * z);
+        stdev = sqrt(0.25 * h * (sinh(z) - z) * (1. - x * x) / (z * z * z));
     }
-    return mean + random_standard_normal(bitgen_state) * sqrt(variance);
+
+    while (n--) {
+        out[n] = mean + random_standard_normal(bitgen_state) * stdev;
+    }
 }
 
 #ifndef PGM_GAMMA_LIMIT
@@ -63,17 +59,24 @@ random_polyagamma_normal_approx(bitgen_t* bitgen_state, double h, double z)
  *
  * The infinite sum is truncated to 200 terms.
  */
-static PGM_INLINE double
-random_polyagamma_gamma_conv(bitgen_t* bitgen_state, double h, double z)
+static PGM_INLINE void
+random_polyagamma_gamma_conv(bitgen_t* bitgen_state, double h, double z,
+                             size_t n, double* out)
 {
     z = 0.5 * fabs(z);
     static const double pi2 = 9.869604401089358;
-    double out = 0., n = 0.5, z2 = z * z;
+    double z2 = z * z;
 
-    do {
-        out += random_standard_gamma(bitgen_state, h) / (pi2 * n * n + z2);
-    } while (++n < PGM_GAMMA_LIMIT);
-    return 0.5 * out;
+    memset(out, 0, n * sizeof(*out));
+
+    while(n--) {
+        double m = 0.5;
+        do {
+            out[n] += random_standard_gamma(bitgen_state, h) / (pi2 * m * m + z2);
+        } while (++m < PGM_GAMMA_LIMIT);
+
+        out[n] *= 0.5;
+    }
 }
 
 /*
@@ -82,25 +85,27 @@ random_polyagamma_gamma_conv(bitgen_t* bitgen_state, double h, double z)
  *
  * Refer to README.md file for more details.
  */
-static PGM_INLINE double
-random_polyagamma_hybrid(bitgen_t* bitgen_state, double h, double z)
+static PGM_INLINE void
+random_polyagamma_hybrid(bitgen_t* bitgen_state, double h, double z,
+                         size_t n, double* out)
 {
     if (h > 50.) {
-        return random_polyagamma_normal_approx(bitgen_state, h, z);
+        random_polyagamma_normal_approx(bitgen_state, h, z, n, out);
     }
     else if (h >= 25. || (((h > 12. && h == (size_t)h) || h >= 8.) && z < 2.)) {
-        return random_polyagamma_saddle(bitgen_state, h, z);
+        random_polyagamma_saddle(bitgen_state, h, z, n, out);
     }
     else if (h == 1. || (h == (size_t)h && z < 2.)) {
-        return random_polyagamma_devroye(bitgen_state, h, z);
+        random_polyagamma_devroye(bitgen_state, h, z, n, out);
     }
     else {
-       return random_polyagamma_alternate(bitgen_state, h, z);
+       random_polyagamma_alternate(bitgen_state, h, z, n, out);
     }
 }
 
 
-typedef double (*pgm_func_t)(bitgen_t* bitgen_state, double h, double z);
+typedef void
+(*pgm_func_t)(bitgen_t* bitgen_state, double h, double z, size_t n, double* out);
 
 const pgm_func_t sampling_method_table[] = {
     [ALTERNATE] = random_polyagamma_alternate,
@@ -114,5 +119,28 @@ const pgm_func_t sampling_method_table[] = {
 PGM_INLINE double
 pgm_random_polyagamma(bitgen_t* bitgen_state, double h, double z, sampler_t method)
 {
-    return sampling_method_table[method](bitgen_state, h, z);
+    double out;
+
+    sampling_method_table[method](bitgen_state, h, z, 1, &out);
+    return out;
+}
+
+
+PGM_INLINE void
+pgm_random_polyagamma_fill(bitgen_t* bitgen_state, double h, double z,
+                           sampler_t method, size_t n, double* out)
+{
+    sampling_method_table[method](bitgen_state, h, z, n, out);
+}
+
+
+PGM_INLINE void
+pgm_random_polyagamma_fill2(bitgen_t* bitgen_state, const double* h, const double* z,
+                            sampler_t method, size_t n, double* restrict out)
+{
+    pgm_func_t f = sampling_method_table[method];
+
+    while (n--) {
+        f(bitgen_state, h[n], z[n], 1, out + n);
+    }
 }