--- a/src/pgm_saddle.h
+++ b/src/pgm_saddle.h
@@ -1,8 +1,11 @@
 #ifndef PGM_SADDLE_H
 #define PGM_SADDLE_H
+#include <stddef.h>
 
 typedef struct bitgen bitgen_t;
 
-double random_polyagamma_saddle(bitgen_t* bitgen_state, double h, double z);
+void
+random_polyagamma_saddle(bitgen_t* bitgen_state, double h, double z,
+                         size_t n, double* out);
 
 #endif