closes #74

To avoid code duplication, I decided in making sure all concrete implementations of the sampling methods accept a size and pointer parameter. After experimentation this turned out to be the most sensible approach given the constraints. As far as speed improvements, it turns out this was a success.

Main branch timings:
```ipynb
In [1]: from polyagamma import random_polyagamma

In [2]: import numpy as np

In [3]: rng = np.random.default_rng(1)

In [4]: %timeit random_polyagamma(20, 5, random_state=rng, size=20000, method='devroye')
50.8 ms ± 704 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [5]: %timeit random_polyagamma(20, 5, random_state=rng, size=20000, method='alternate')
26 ms ± 710 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [6]: %timeit random_polyagamma(20, 5, random_state=rng, size=20000, method='saddle')
29.9 ms ± 519 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [7]: %timeit random_polyagamma(20, 5, random_state=rng, size=20000, method='gamma')
173 ms ± 2.57 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

```
This branch's timings:
```ipynb
In [1]: from polyagamma import random_polyagamma

In [2]: import numpy as np

In [3]: rng = np.random.default_rng(1)

In [4]: %timeit random_polyagamma(20, 5, random_state=rng, size=20000, method='devroye')
48.3 ms ± 32.8 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [5]: %timeit random_polyagamma(20, 5, random_state=rng, size=20000, method='alternate')
13.9 ms ± 24.9 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [6]: %timeit random_polyagamma(20, 5, random_state=rng, size=20000, method='saddle')
8.97 ms ± 55.6 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [7]: %timeit random_polyagamma(20, 5, random_state=rng, size=20000, method='gamma')
159 ms ± 196 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)

```

`devroye`: 4.92% boost
`alternate`: 87% boost
`saddle`: 233% boost
`gamma`: 8.8% boost

The speed boost using the default parameter values (`h=1`, `z=0`) are:

`devroye`: 8.67% boost
`alternate`: 50% boost
`saddle`: 159% boost
`gamma`: 0% boost